marchukov/project_03/divolte-unpack/requirements.txt
kafka
lz4
avro-python3
